parse user agent correctly
